Move classes to its files
