+++ Ctoss/ICache.cs
@@ -1,0 +1,7 @@
+namespace Ctoss;
+
+public interface ICache<in TKey, TEntry>
+{
+    Task<object>? Get(TKey key);
+    Task Set(TKey key, TEntry entry);
+}
