+++ Ctoss/ICtossService.cs
@@ -1,0 +1,8 @@
+using Ctoss.Models.AgGrid;
+
+namespace Ctoss;
+
+public interface ICtossService
+{
+    Task<AgGridQueryResult<T>> ApplyAsync<T>(AgGridQuery query, IEnumerable<T> enumerable);
+}
