+++ Ctoss/IFilterExpressionCache.cs
@@ -1,0 +1,7 @@
+using Ctoss.Models;
+
+namespace Ctoss;
+
+public interface IFilterExpressionCache : ICache<FilterDescriptor, object>
+{
+}
